chore: refactor code sections
