+++ src/index.ts
@@ -1,42 +1,16 @@
 import * as core from '@actions/core';
-import * as github from '@actions/github';
-import {
-  ChannelType,
-  Client,
-  Message,
-  ThreadAutoArchiveDuration,
-  ThreadChannel,
-} from 'discord.js';
 
-import {getInputs, Inputs, setOutput} from './io';
-import {getIssueCommentLink, getThreadLink, replaceThreadLink} from './utils';
+import {createThread, initClient} from './discord';
+import {commentOnIssue} from './github';
+import {getInputs, setOutput} from './io';
+import {getIssueCommentLink, getThreadLink} from './utils';
 
 (async () => {
   core.startGroup('Checking inputs...');
   const inputs = await getInputs();
   core.endGroup();
 
-  return initClient(inputs);
-})().catch(e => {
-  core.endGroup();
-  core.setFailed(e);
-});
-
-/**
- * Creates the Discord.js client
- * @param inputs The action inputs
- */
-async function initClient(inputs: Inputs) {
-  core.startGroup('Initializing Discord client...');
-
-  const client = new Client({
-    intents: ['Guilds'],
-  });
-
-  client.on('error', core.setFailed);
-  client.on('warn', core.warning);
-
-  client.on('ready', async client => {
+  return initClient(inputs, async client => {
     try {
       const thread = await createThread(client, inputs);
 
@@ -69,82 +43,7 @@ async function initClient(inputs: Inputs) {
       process.exit();
     }
   });
-
-  client.login(inputs.discord_bot_token);
-
+})().catch(e => {
   core.endGroup();
-}
-
-/**
- * Creates a Discord thread in the specified channel
- * @param client The Discord.js Client
- * @param inputs The actions inputs
- * @returns The ThreadChannel instance of the created thread
- */
-async function createThread(
-  client: Client,
-  inputs: Inputs
-): Promise<ThreadChannel> {
-  core.startGroup('Creating thread...');
-  core.info(`Logged in as ${client.user?.tag}`);
-
-  const channel = await client.channels.fetch(inputs.destination.channel);
-  if (!channel || channel.type !== ChannelType.GuildText)
-    throw new Error(`${channel?.id} is not a text channel`);
-
-  let message: Message;
-  if (inputs.destination.message) {
-    message = await channel.messages.fetch(inputs.destination.message);
-    if (!message)
-      throw new Error(`Message ${inputs.destination.message} not found`);
-  } else {
-    message = await channel.send(inputs.discord_message);
-  }
-
-  const thread = await message.startThread({
-    name: `#${inputs.issue.number} - ${inputs.issue.title}`,
-    autoArchiveDuration: ThreadAutoArchiveDuration.OneWeek,
-  });
-
-  core.info(`Created thread ${thread.id}`);
-  core.endGroup();
-
-  return thread;
-}
-
-/**
- * Comments on the original issue with the thread link
- * @param inputs The action inputs
- * @param threadLink The link to the thread
- * @returns The ID of the comment if posted, `null` if disabled in inputs
- */
-async function commentOnIssue(
-  inputs: Inputs,
-  threadLink: string
-): Promise<number | null> {
-  if (!inputs.issue_comment) {
-    core.info('Not going to comment on issue.');
-    return null;
-  }
-
-  core.startGroup('Commenting on issue...');
-  const octokit = github.getOctokit(inputs.github_token);
-
-  const comment_tag = inputs.issue_comment_tag
-    ? '\n\n<sup>🤖 This comment was automatically posted by <a href="https://github.com/EndBug/discuss-on-discord">Discuss on Discord</a></sup>'
-    : '';
-
-  const res = await octokit.rest.issues.createComment({
-    owner: inputs.issue.repo.owner,
-    repo: inputs.issue.repo.name,
-    issue_number: inputs.issue.number,
-    body: replaceThreadLink(inputs.issue_comment, threadLink) + comment_tag,
-  });
-
-  const commentID = res.data.id;
-
-  core.info(`Commented on issue with comment ${commentID}`);
-  core.endGroup();
-
-  return commentID;
-}
+  core.setFailed(e);
+});
